Split GA pipelines to prevent issue
