+++ .github/workflows/python-package.yml
@@ -1,11 +1,34 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Python package
+name: Test repositories
 on: [push, pull_request]
 
 jobs:
-  build:
+  test-repositories:
+
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install .
+        python -m pip install pytest pytest-xdist
+    - name: Test with pytest
+      run: |
+        pytest -n 4 --ignore=tests/test_repositories_plus.py
+
+  test-repositories-plus:
 
     runs-on: ubuntu-latest
     strategy:
@@ -26,4 +49,4 @@ jobs:
         python -m pip install pytest pytest-xdist
     - name: Test with pytest
       run: |
-        pytest -n 4
+        pytest -n 4 tests/test_repositories_plus.py
